server/requirements.txt
# InMemory Server Requirements
# Clean FastAPI server following mem0 pattern

# Core InMemory library (install from parent directory)
# Run: pip install -e .. 

# FastAPI server requirements
fastapi>=0.100.0
uvicorn>=0.20.0
python-dotenv>=1.0.0

# HTTP client for potential external calls
httpx>=0.25.0

# Development dependencies (optional)
pytest>=7.0.0
pytest-asyncio>=0.21.0
